requirements.txt
alembic==1.10.4
bcrypt==4.0.1
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
Deprecated==1.2.13
Flask==2.3.1
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.4.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
Pygments==2.15.1
PyJWT==2.6.0
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.29.0
rich==13.3.5
SQLAlchemy==2.0.11
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
validators==0.20.0
Werkzeug==2.3.2
wrapt==1.15.0
yapf==0.33.0
